+++ docs/Examples.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Row, Col } from 'rsuite';
+import { Row, Col, Nav } from 'rsuite';
 import CodeView from 'react-code-view';
 
 const CustomCodeView = ({ ...props }) => (
@@ -10,16 +10,42 @@ const CustomCodeView = ({ ...props }) => (
   </Row>
 );
 
-const Examples = ({ list, dependencies }) => {
-  return (
-    <div>
-      {list.map((item, index) => (
+class Examples extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      index: 0
+    };
+  }
+  render() {
+    const { list, dependencies } = this.props;
+    const { index } = this.state;
+    console.log(index);
+
+    return (
+      <div>
+        <Nav tabs>
+          {list.map((item, i) => {
+            return (
+              <Nav.Item
+                eventKey={i}
+                active={index === i}
+                key={i}
+                onClick={() => {
+                  this.setState({ index: i });
+                }}
+              >
+                {item.title}
+              </Nav.Item>
+            );
+          })}
+        </Nav>
         <CustomCodeView key={index} dependencies={dependencies}>
-          {item}
+          {list[index].content}
         </CustomCodeView>
-      ))}
-    </div>
-  );
-};
+      </div>
+    );
+  }
+}
 
 export default Examples;