+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 > 2018-6-26
 
-- Feature: Support `loadAnimaction` on `<Table>` ([#27])
+- Feature: Support `loadAnimation` on `<Table>` ([#27])
 - Bugfix: The position was updated incorrectly when the scrollbar was clicked on `<Table>` ([#26])
 
 [27]: https://github.com/rsuite/rsuite-table/pull/27